Bugfix: Re-add TrainInput.cs to prefab containing input scripts
This scene uses the prefab for input scripts, but was modified a little by adding the Train Input script as well for demonstrative purposes, which is usually bound to the Train object. This prefab was automatically updated for some reason, so the Train Input script was removed. This commit fixes that small slip-up by adding it back and binding the sliders to it.